app/requirements.txt
Django>=2.2.4,<2.5.0
djangorestframework>=3.9.0,<3.10.0
gunicorn>=20.0.4,<20.1.0
psycopg2-binary>=2.8.5,<2.9.0
flake8>=3.6.0,<3.7.0
ptvsd>=4.3.2,<4.4.0
Pillow>=7.1.2,<7.2.0